Log, but do not kill when the memory score is untouched

+++ buildkitd/entrypoint.sh
@@ -267,8 +267,12 @@ do
         # common instance is when something is OOM killed, for instance. This enumerates
         # all those PIDs, and kills them to prevent accidential "ghost" loads.
         if [ "$PID" != "$execpid" ]; then
-            ! "$BUILDKIT_DEBUG" || echo "$(date) | $PID($(cat /proc/"$PID"/cmdline)) killed" >> /var/log/oom_adj
-            kill -9 "$PID"
+            if [ "$OOM_SCORE_ADJ" -ne "0" ]
+                ! "$BUILDKIT_DEBUG" || echo "$(date) | $PID($(cat /proc/"$PID"/cmdline)) killed" >> /var/log/oom_adj
+                kill -9 "$PID"
+            else 
+                ! "$BUILDKIT_DEBUG" || echo "$(date) | $PID($(cat /proc/"$PID"/cmdline)) was not killed because OOM_SCORE_ADJ was default or not set" >> /var/log/oom_adj
+            fi
         fi
     done
